Try to ensure we got subject.txt alright.
For reasons as yet unknown, sometimes dis only returns partial responses.
When this happens, the size of the response we got always seems to be a
multiple of 64; I don't if that's meaningful. I also don't know if this is
a problem with CloudFlare or requests or what.

The JSON-parsing step already ensured that partial responses got retried in
the case of posts. Now we're also checking if subject.txt looks alright.
It's not ideal, but since dis doesn't specify Content-Length in the header,
it's the only thing that came to mind.